requirements.txt
fastapi==0.112.2
uvicorn==0.30.6
PyYAML==6.0.2